pyproject.toml
[project]
name = "student-retention-platform"
version = "1.0.0"
description = "AI-driven Student Retention Platform - B2B SaaS for Universities"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
    {name = "Student Retention Platform Team"}
]
keywords = ["education", "machine-learning", "student-retention", "dropout-prediction"]

dependencies = [
    # Core
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    
    # Database
    "sqlalchemy>=2.0.25",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    
    # Authentication
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "authlib>=1.3.0",
    
    # ML & Data
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "shap>=0.44.0",
    "fairlearn>=0.9.0",
    "interpret>=0.5.0",
    "joblib>=1.3.0",
    
    # Async & Background Tasks
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "aiofiles>=23.2.0",
    
    # Monitoring & Logging
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "opentelemetry-api>=1.22.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "tenacity>=8.2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
